Try to revert packages to a stable version
Upgrading contact-form-7 to 5.3.2 and the other changes I made broke the
site. I'm trying to get it back to a stable version so I can update it
one change at a time, in the hopes to make it work.

I've removed a plugin for local dev that is not necessary.